+++ rss/models/__init__.py
@@ -1,0 +1,3 @@
+from .torrent import Torrent  # noqa
+from .transmission_client import TransmissionClient  # noqa
+from .feed import Feed  # noqa
