+++ rss/models/feed.py
@@ -1,0 +1,28 @@
+from django.db import models
+
+from rss.fields import RegexField
+
+
+class Feed(models.Model):
+    url = models.URLField()
+    regex_filter = RegexField(max_length=2048, blank=True)
+    expires_at = models.DateTimeField(
+        blank=True, null=True, help_text="Skip feed fetch after this date"
+    )
+    download_dir = models.CharField(max_length=2048, blank=True)
+    ignore_older_than = models.DateTimeField(
+        blank=True, null=True, help_text="Ignore older torrents than this date"
+    )
+    ignore_newer_than = models.DateTimeField(
+        blank=True, null=True, help_text="Ignore newer torrents than this date"
+    )
+    start_paused = models.BooleanField(default=False, help_text="Start torrents paused")
+    chronological = models.BooleanField(
+        default=True, help_text="Add torrents in chronological order"
+    )
+    transmission_clients = models.ManyToManyField(
+        to="rss.TransmissionClient", related_name="feeds", blank=True
+    )
+
+    def __str__(self):
+        return f"{self.url}"
